fix more problems, add validation on file type and roles users, refactoring code in S3 service

+++ src/main/java/faang/school/projectservice/service/resource/ResourceService.java
@@ -1,14 +1,18 @@
 package faang.school.projectservice.service.resource;
 
+import faang.school.projectservice.exception.AccessDeniedException;
+import faang.school.projectservice.exception.DataNotFoundException;
 import faang.school.projectservice.exception.FileException;
-import faang.school.projectservice.exception.ResourceNotFoundException;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.Resource;
 import faang.school.projectservice.model.ResourceStatus;
+import faang.school.projectservice.model.ResourceType;
 import faang.school.projectservice.model.TeamMember;
+import faang.school.projectservice.model.TeamRole;
 import faang.school.projectservice.repository.ProjectRepository;
 import faang.school.projectservice.repository.ResourceRepository;
 import faang.school.projectservice.repository.TeamMemberRepository;
+import faang.school.projectservice.service.s3.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,9 +24,11 @@ import java.util.List;
 import java.util.Optional;
 
 import static faang.school.projectservice.service.resource.ResourceErrorMessage.FILE_SIZE_EXCEED_STORAGE_VALUE;
+import static faang.school.projectservice.service.resource.ResourceErrorMessage.INVALID_FILE_EXTENSION;
 import static faang.school.projectservice.service.resource.ResourceErrorMessage.NOT_FOUND_PROJECT;
 import static faang.school.projectservice.service.resource.ResourceErrorMessage.NOT_FOUND_RESOURCE;
 import static faang.school.projectservice.service.resource.ResourceErrorMessage.NOT_FOUND_TEAM_MEMBER;
+import static faang.school.projectservice.service.resource.ResourceErrorMessage.NOT_PERMISSION_TO_DELETE;
 
 
 @RequiredArgsConstructor
@@ -40,6 +46,12 @@ public class ResourceService {
     public Resource addResource(Long projectId, Long userId, MultipartFile file) {
         Project project = getProjectById(projectId);
         TeamMember member = getTeamMember(userId, projectId);
+        ResourceType fileContentType = ResourceType.getResourceType(file.getContentType());
+
+        if (fileContentType == ResourceType.NONE ||
+                fileContentType == ResourceType.OTHER) {
+            throw new FileException(INVALID_FILE_EXTENSION);
+        }
 
         String folder = String.format(folderPattern, projectId, project.getName());
         Resource resource = s3Service.uploadFile(file, folder);
@@ -59,6 +71,12 @@ public class ResourceService {
         Project project = getProjectById(projectId);
         TeamMember member = getTeamMember(userId, projectId);
         Resource oldResource = getResourceById(resourceId);
+
+        ResourceType fileContentType = ResourceType.getResourceType(file.getContentType());
+
+        if (oldResource.getType() != fileContentType) {
+            throw new FileException(INVALID_FILE_EXTENSION);
+        }
 
         long sizeDifference = file.getSize() - oldResource.getSize().longValue();
         String folder = String.format(folderPattern, projectId, project.getName());
@@ -82,14 +100,21 @@ public class ResourceService {
         TeamMember member = getTeamMember(userId, projectId);
         Resource resource = getResourceById(resourceId);
 
+        if (!(member.getRoles().contains(TeamRole.OWNER) ||
+                member.getRoles().contains(TeamRole.MANAGER))) {
+            throw new AccessDeniedException(NOT_PERMISSION_TO_DELETE);
+        }
+
         s3Service.deleteFile(resource.getKey());
 
+        setNewStorageSize(project, resource.getSize().negate().longValue());
+
+        resource.setKey(null);
+        resource.setSize(null);
         resource.setStatus(ResourceStatus.DELETED);
         resource.setUpdatedAt(LocalDateTime.now());
         resource.setUpdatedBy(member);
         resourceRepository.save(resource);
-
-        setNewStorageSize(project, resource.getSize().negate().longValue());
     }
 
     private void setNewStorageSize(Project project, Long fileSize) {
@@ -107,16 +132,16 @@ public class ResourceService {
 
     private Project getProjectById(Long projectId) {
         return projectRepository.findById(projectId)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format(NOT_FOUND_PROJECT, projectId)));
+                .orElseThrow(() -> new DataNotFoundException(String.format(NOT_FOUND_PROJECT, projectId)));
     }
 
     private TeamMember getTeamMember(Long userId, Long projectId) {
         return Optional.ofNullable(teamMemberRepository.findByUserIdAndProjectId(userId, projectId))
-                .orElseThrow(() -> new ResourceNotFoundException(String.format(NOT_FOUND_TEAM_MEMBER, userId)));
+                .orElseThrow(() -> new DataNotFoundException(String.format(NOT_FOUND_TEAM_MEMBER, userId)));
     }
 
     private Resource getResourceById(Long resourceId) {
         return resourceRepository.findById(resourceId)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format(NOT_FOUND_RESOURCE, resourceId)));
+                .orElseThrow(() -> new DataNotFoundException(String.format(NOT_FOUND_RESOURCE, resourceId)));
     }
 }
